better thermo precision

+++ lib/Common/heat_state.h
@@ -7,14 +7,14 @@ struct HeatState {
 
   bool brewActive;                          // Only when brewing, thermoCompensateHeat is caculated, otherwize only uses pidAdjustHeat
   bool lastBoilerState;                     // false is OFF, true is ON
-  uint32_t lastBoilerStateTimestamp;        // mills
+  uint32_t lastBoilerStateTimestamp;        // micros for better accurcy, for better RTD Sensor pt100 & ad7124 ADC.
   float lastWaterPumped;                    // ml or g
   uint32_t lastWaterPumpedTimestamp;        // mills
-  float heatBalancePool;                    // Unit is Joule == Watt*Second
-  uint32_t lastPidOutputTimestamp;          // mills
-  float pidOutput;                          // pid out, percentage of full power, no need to reset when brew on-off
+  double heatBalancePool;                   // Unit is Joule == Watt*Second
+  uint32_t lastPidOutputTimestamp;          // micros for better accurcy
+  double pidOutput;                         // pid out, percentage of full power, no need to reset when brew on-off
   float lastTemperature;                    // last Sensor read temperature
-  float lastTemperatureTime;                // last Sensor read temperature timw stamp
+  float lastTemperatureTime;                // last Sensor read temperature timestamp
   float lastThermoCompensateHeat;           // Thermo Compensation heat caculate, Unit is Joule
   float lastThermoHeaterConsumed;           // Unit is Joule
   bool isBoilerOperatorTC;                  // Who operated the boiler, true is Thermo Compensation, false is PID or Hard Limit.
